+++ src/org/foxteam/noisyfox/widget/PullHeaderScrollView.java
@@ -137,7 +137,7 @@ public class PullHeaderScrollView extends ScrollView implements IHeaderPull {
 	@Override
 	public void pull(int height) {
 		if (mOnPullListener != null) {
-			float p = 100.0F * ((height - mHeaderHeightMin) / (mHeaderHeightMax - mHeaderHeightMin));
+			float p = 100.0F * ((float) (height - mHeaderHeightMin) / (float) (mHeaderHeightMax - mHeaderHeightMin));
 			mOnPullListener.onPull((int) (p / 0.7F));
 		}
 	}
